change report Charge to uint64

+++ opendsp/report.go
@@ -5,7 +5,7 @@ type CampaignReport struct {
 	DspCustId  string `json:"dsp_cust_id,omitempty" codec:",omitempty"` // dsp系统中的推广主id
 	CampaignId uint64 `json:"campaign_id,omitempty" codec:",omitempty"` // 推广计划id
 	RecordOn   string `json:"thedate,omitempty" codec:",omitempty"`     // 日期
-	Charge     uint   `json:"charge,omitempty" codec:",omitempty"`      // 消耗
+	Charge     uint64 `json:"charge,omitempty" codec:",omitempty"`      // 消耗
 	Click      uint   `json:"click,omitempty" codec:",omitempty"`       // 点击
 	Pv         uint64 `json:"pv,omitempty" codec:",omitempty"`          // 展现量
 }
@@ -15,7 +15,7 @@ type AdgroupReport struct {
 	DspCustId string `json:"dsp_cust_id,omitempty" codec:",omitempty"` // dsp系统中的推广主id
 	AdgroupId uint64 `json:"adgroup_id,omitempty" codec:",omitempty"`  // 推广单元id
 	RecordOn  string `json:"thedate,omitempty" codec:",omitempty"`     // 日期
-	Charge    uint   `json:"charge,omitempty" codec:",omitempty"`      // 消耗
+	Charge    uint64 `json:"charge,omitempty" codec:",omitempty"`      // 消耗
 	Click     uint   `json:"click,omitempty" codec:",omitempty"`       // 点击
 	Pv        uint64 `json:"pv,omitempty" codec:",omitempty"`          // 展现量
 }
@@ -25,7 +25,7 @@ type CreativeReport struct {
 	DspCustId  string `json:"dsp_cust_id,omitempty" codec:",omitempty"` // dsp系统中的推广主id
 	CreativeId uint64 `json:"creative_id,omitempty" codec:",omitempty"` // 创意id
 	RecordOn   string `json:"thedate,omitempty" codec:",omitempty"`     // 日期
-	Charge     uint   `json:"charge,omitempty" codec:",omitempty"`      // 消耗
+	Charge     uint64 `json:"charge,omitempty" codec:",omitempty"`      // 消耗
 	Click      uint   `json:"click,omitempty" codec:",omitempty"`       // 点击
 	Pv         uint64 `json:"pv,omitempty" codec:",omitempty"`          // 展现量
 }
@@ -35,7 +35,7 @@ type CrowdReport struct {
 	DspCustId string `json:"dsp_cust_id,omitempty" codec:",omitempty"` // dsp系统中的推广主id
 	CrowdId   uint64 `json:"crowd_id,omitempty" codec:",omitempty"`    // 人群id
 	RecordOn  string `json:"thedate,omitempty" codec:",omitempty"`     // 日期
-	Charge    uint   `json:"charge,omitempty" codec:",omitempty"`      // 消耗
+	Charge    uint64 `json:"charge,omitempty" codec:",omitempty"`      // 消耗
 	Click     uint   `json:"click,omitempty" codec:",omitempty"`       // 点击
 	Pv        uint64 `json:"pv,omitempty" codec:",omitempty"`          // 展现量
 }
